refactor: PasswordUtil의 메인 메서드를 따로 클래스로 만들어 분리
